Apply OS/2-EMX specific files and patches.
Note: This is a cumulative patch containing the work of several people
which is impossible to separate now.

This patch contains a number of '#ifdef 0'-like hacks that need to be
reviewed and fixed properly as they are unlikely to be accepted upstream
in the current form. Also, some patches may be not relevant any more.
(cherry picked from commit 25ebd257ce836444e0914e35a7baf3fdff67e3b9)

Conflicts:

        Makefile.in
        gcc/Makefile.in
        gcc/config.build
        gcc/config.gcc
        gcc/config.host
        gcc/config/i386/i386-protos.h
        gcc/config/i386/i386.c
        gcc/config/i386/i386.h
        gcc/config/i386/i386.md
        gcc/cp/class.c
        gcc/cp/decl.c
        gcc/cp/error.c
        gcc/cp/g++spec.c
        gcc/cp/mangle.c
        gcc/cppdefault.c
        gcc/fix-header.c
        gcc/gcc.c
        gcc/gengtype.c
        gcc/gthr.h
        gcc/langhooks.c
        gcc/libgcc2.c
        gcc/target-def.h
        include/filenames.h
        libgcc/config.host
        libiberty/configure
        libiberty/configure.ac
        libstdc++-v3/include/bits/stringfwd.h
        libstdc++-v3/libsupc++/typeinfo