Apply OS/2-EMX specific files and patches.
Note: This is a cumulative patch containing the work of several people
which is impossible to separate now.

This patch contains a number of '#ifdef 0'-like hacks that need to be
reviewed and fixed properly as they are unlikely to be accepted upstream
in the current form. Also, some patches may be not relevant any more.