Remove version number from system import library for gccXXX.dll.
The library is now named libgcc_so_d which corresponds to the
original (old) behavior. The comments said that it was done to solve
bootstrapping problems but since we configure with --disable-bootstrap
now, it's irrelevant. More over, I'm not really sure that using libgcc
funcitons from GCC 3 (that live in libcXXX.dll) in GCC 4 while
bootstrapping was correct anyway.