EMX: Fix regression in 592f328.
This finally enables code from 667cbd1 but w/o the #ifdef __EMX__ clause -
the fix enclosed there should be cross-platform: a NULL .language for an
input file means it IS a source or an object file (for which the -x option
was simply not given on the command line, i.e. the usual case) and therefore
it should not be skipped but used as a base for the output file name.