OS/2: Define max_align_t on C++11
C++11 requires max_align_t which is a C11 stuff. However, OS/2 kLIBC
headers are not ready for C11.

So, define max_align_t in os_defines.h which is a OS-specefic header
on C++. By this, it's possible to remove _GCC_MAX_ALIGN_T guard of
cstddef, too.

    modified:   libstdc++-v3/config/os/os2/os_defines.h
    modified:   libstdc++-v3/include/c_global/cstddef

Close #26.